Add files via upload
Still working on the basis, even do this is all just for test, i can't just throw trash code everywhere and hope it all work in the end... i need to have a solid foundation none the less. __Solid__ foundation, I mean i know it's full of bad coding practice but i have only one goal, and that is something that work and can be played as soon as possible.